+++ sql/expression/function/convert_tz.go
@@ -15,17 +15,11 @@
 package function
 
 import (
-	"errors"
 	"fmt"
-	"math"
-	"regexp"
-	"time"
-
+	gmstime "github.com/dolthub/go-mysql-server/internal/time"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/types"
 )
-
-var offsetRegex = regexp.MustCompile(`(?m)^([+\-])(\d{2}):(\d{2})$`) // (?m)^\+|\-(\d{2}):(\d{2})$
 
 type ConvertTz struct {
 	dt     sql.Expression
@@ -110,11 +104,8 @@ func (c *ConvertTz) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
 		return nil, nil
 	}
 
-	t := time.Now()
-	_, offset := t.Zone()
-
 	if fromStr == globalTimeZone.Default {
-		fromStr = getSystemDelta(offset)
+		fromStr = gmstime.SystemDelta()
 	}
 
 	toStr, ok := to.(string)
@@ -123,91 +114,15 @@ func (c *ConvertTz) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
 	}
 
 	if toStr == globalTimeZone.Default {
-		toStr = getSystemDelta(offset)
+		toStr = gmstime.SystemDelta()
 	}
 
-	converted, success := convertTimeZone(datetime, fromStr, toStr)
-	if success {
-		return types.Datetime.ConvertWithoutRangeCheck(converted)
-	}
-
-	// If we weren't successful converting by timezone try converting via offsets.
-	converted, success = convertOffsets(datetime, fromStr, toStr)
+	converted, success := gmstime.ConvertTimeZone(datetime, fromStr, toStr)
 	if !success {
 		return nil, nil
 	}
 
 	return types.Datetime.ConvertWithoutRangeCheck(converted)
-}
-
-// convertTimeZone returns the conversion of t from timezone fromLocation to toLocation.
-func convertTimeZone(datetime time.Time, fromLocation string, toLocation string) (time.Time, bool) {
-	fLoc, err := time.LoadLocation(fromLocation)
-	if err != nil {
-		return time.Time{}, false
-	}
-
-	tLoc, err := time.LoadLocation(toLocation)
-	if err != nil {
-		return time.Time{}, false
-	}
-
-	delta := getCopy(datetime, fLoc).Sub(getCopy(datetime, tLoc))
-
-	return datetime.Add(delta), true
-}
-
-// getCopy recreates the time t in the wanted timezone.
-func getCopy(t time.Time, loc *time.Location) time.Time {
-	return time.Date(t.Year(), t.Month(), t.Day(), t.Hour(), t.Minute(), t.Second(), t.Nanosecond(), loc).UTC()
-}
-
-// convertOffsets returns the conversion of t to t + (endDuration - startDuration) and a boolean indicating success.
-func convertOffsets(t time.Time, startDuration string, endDuration string) (time.Time, bool) {
-	fromDuration, err := getDeltaAsDuration(startDuration)
-	if err != nil {
-		return time.Time{}, false
-	}
-
-	toDuration, err := getDeltaAsDuration(endDuration)
-	if err != nil {
-		return time.Time{}, false
-	}
-
-	return t.Add(toDuration - fromDuration), true
-}
-
-// getDeltaAsDuration takes in a MySQL offset in the format (ex +01:00) and returns it as a time Duration.
-func getDeltaAsDuration(d string) (time.Duration, error) {
-	var hours string
-	var mins string
-	var symbol string
-	matches := offsetRegex.FindStringSubmatch(d)
-	if len(matches) == 4 {
-		symbol = matches[1]
-		hours = matches[2]
-		mins = matches[3]
-	} else {
-		return -1, errors.New("error: unable to process time")
-	}
-
-	return time.ParseDuration(symbol + hours + "h" + mins + "m")
-}
-
-func getSystemDelta(offset int) string {
-	seconds := offset % (60 * 60 * 24)
-	hours := math.Floor(float64(seconds) / 60 / 60)
-	seconds = offset % (60 * 60)
-	minutes := math.Floor(float64(seconds) / 60)
-
-	result := fmt.Sprintf("%02d:%02d", int(math.Abs(hours)), int(math.Abs(minutes)))
-	if offset >= 0 {
-		result = fmt.Sprintf("+%s", result)
-	} else {
-		result = fmt.Sprintf("-%s", result)
-	}
-
-	return result
 }
 
 // Children implements the sql.Expression interface.
